Crate lib_rust_primes

Not too terrible Prime Calculator

written in Rust with Python bindings

See Python docs for use in Python.

Sample output and runtime on an Apple M1 Pro:

Calculating all primes under 10^4:
- SieveMethod.ATKIN                       : 0.000024s
- SieveMethod.ERATOSTHENES                : 0.000011s
- SieveMethod.ERATOSTHENES_THREADED       : 0.000142s
- SieveMethod.ERATOSTHENES_ATOMIC         : 0.000115s

Calculating all primes under 10^5:
- SieveMethod.ATKIN                       : 0.000190s
- SieveMethod.ERATOSTHENES                : 0.000099s
- SieveMethod.ERATOSTHENES_THREADED       : 0.000221s
- SieveMethod.ERATOSTHENES_ATOMIC         : 0.000288s

Calculating all primes under 10^6:
- SieveMethod.ATKIN                       : 0.001648s
- SieveMethod.ERATOSTHENES                : 0.001109s
- SieveMethod.ERATOSTHENES_THREADED       : 0.001364s
- SieveMethod.ERATOSTHENES_ATOMIC         : 0.001368s

Calculating all primes under 10^7:
- SieveMethod.ATKIN                       : 0.014422s
- SieveMethod.ERATOSTHENES                : 0.011521s
- SieveMethod.ERATOSTHENES_THREADED       : 0.024931s
- SieveMethod.ERATOSTHENES_ATOMIC         : 0.011437s

Calculating all primes under 10^8:
- SieveMethod.ATKIN                       : 0.447571s
- SieveMethod.ERATOSTHENES                : 0.423881s
- SieveMethod.ERATOSTHENES_THREADED       : 0.222338s
- SieveMethod.ERATOSTHENES_ATOMIC         : 0.153570s

Calculating all primes under 10^9:
- SieveMethod.ATKIN                       : 5.407987s
- SieveMethod.ERATOSTHENES                : 5.830421s
- SieveMethod.ERATOSTHENES_THREADED       : 2.663690s
- SieveMethod.ERATOSTHENES_ATOMIC         : 2.096283s

Modules

prelude

A module containing all the exports of this library.

Structs

IterAtomicSievePrimes

An Iterator struct to yield the lowest prime currently found in the sieve, up to the square root of the length of the sieve.

Traits

CanBeTrue

Trait shared by bool and AtomicBool for global implementation.

CanSieve

A trait for SieveMethod enum as well as its variants, allowing for the unified CanSieve::sieve() method to be implemented. This method is then used to implement the downstream methods of CanSieve::list_primes(), CanSieve::count_primes() and CanSieve::is_prime().

IsAtomicSieve

Special trait for AtomicSieves, which

SievedArray

Trait allowing Arrays of bool can be collected into Vec<u64> after sieving.

Type Definitions

AtomicSieve

NonAtomicSieve

OwnedSieve

Sieve

ViewSieve