Enum lib_rust_primes::prelude::SieveMethod

pub enum SieveMethod {
    ATKIN,
    ERATOSTHENES,
    ERATOSTHENES_ATOMIC,
    ERATOSTHENES_THREADED,
}

Methods of sieving.

To be used with :func:~rust_primes.count_primes

Variants

ATKIN

Modern method, but less well optimised by the compiler; not necessarily more performant.

ERATOSTHENES

The ancient method. Using the :meth:ndarray.slice_mut().step method, the compiler can optimise the inner loop to a close to O(n) operation.

ERATOSTHENES_ATOMIC

Same method as :attr:ERATOSTHENES, but conducted on an array of Atomic booleans instead. Allows threading to run withoutmuch overhead. This is the default.

ERATOSTHENES_THREADED

Experimental. An attempt to introduce threading into :attr:ERATOSTHENES. It works by using :attr:ERATOSTHENES to create a base array of primes, upto :func:Math.sqrt of the upper bound, then split the list of primes into equal size for each worker, then each worker sieve the rest of the (larger) numbers.